pyproject.toml
[project]
name = "pingu-bot"
version = "0.1.0"
description = "A fastapi-based LINE Bot that echoes user messages. The bot receives text messages from LINE users and responds by sending back the same message. Built with LINE Bot SDK and FastAPI web framework."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "line-bot-sdk>=3.17.1",
    "pydantic-settings>=2.10.1",
    "structlog>=25.4.0",
]
